fix systemd service plugin
The plugin had a bug, as masking did not work at all, but rather crashed
the app. This commit fixes that, by correcting the function signature
and ensuring the daemon is reloaded after the action. The reloading is
done for all commands now, which might be overkill for some actions
taken, but not harmful.

The commit also ensures that the mask/unmask icons are valid, at least
for the elementary icons. We don't have our own icons, so we use
a "best effort" approach.